requirements.txt
slack-sdk==3.26.2
slack-bolt==1.18.1
python-dotenv==1.0.0
PyGithub==2.1.1
gitpython==3.1.40

